+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `arrCount` method
 - Add `between` method
 - Add `arrGroupByMaxCount` method
+- Add `numRandom` method
 
 ## 1.0.0 (2024-10-19)