+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 1.0.1 (2024--)
+## 1.1.0 (2024-11-26)
 
 - Add `arrCount` method
 - Add `between` method